stub out header data import
