--- config.js
@@ -1,8 +1,0 @@
-module.exports = {
-  baseUrl: process.env.NEXT_PUBLIC_BASE_URL,
-  drupalBaseUrl: process.env.NEXT_PUBLIC_DRUPAL_BASE_URL,
-  defaultLocale: 'en',
-  locales: {
-    en: 'English',
-  },
-}
