+++ src/lib/constants/index.ts
@@ -1,25 +1,41 @@
-export const HEADER_FOOTER_PATH =
-  'https://www.va.gov/generated/headerFooter.json'
-
-export const S3_URL =
-  'https://s3-us-gov-west-1.amazonaws.com/content.www.va.gov'
-
-export const DEV_PATH = 'https://dev.cms.va.gov'
-export const IMAGE_PATH_TEMP = 'https://www.va.gov'
-
-// Build Page nodes: params
-export const FIELDS =
-  'field_answer,field_alert_single.field_alert_non_reusable_ref.field_va_paragraphs,field_buttons,field_tags.field_audience_beneficiares,field_related_information,field_related_benefit_hubs,field_contact_information.field_contact_default'
-
-export const VA_WRAPPER_IMAGES = {
-  logo: `${S3_URL}/img/header-logo.png`,
-  footer: 'https://www.va.gov/img/homepage/va-logo-white.png',
-  usFlag: `${S3_URL}/img/tiny-usa-flag.png`,
-  httpsIcon: `${S3_URL}/img/icon-https.svg`,
-  arrowIcon: `${S3_URL}/img/arrow-right-white.svg`,
-  dotIcon: `${S3_URL}/img/icon-dot-gov.svg`,
-  metaImg: `https://www.va.gov/img/design/logo/va-og-image.png`,
-}
-// content
-export const US_FLAG_MSG =
-  ' An official website of the United States government'
+// These footer links are also hard-coded in src/platform/static-data/footer-links.json
+export const FOOTER_LINKS = [
+  {
+    "column": 4,
+    "href": "https://www.va.gov/resources/",
+    "order": 1,
+    "target": null,
+    "title": "Resources and support"
+  },
+  {
+    "column": 4,
+    "href": "https://www.va.gov/contact-us/",
+    "order": 2,
+    "target": null,
+    "title": "Contact us"
+  },
+  {
+    "column": 4,
+    "label": "Call us",
+    "href": "tel:18006982411",
+    "order": 3,
+    "target": null,
+    "title": "800-698-2411"
+  },
+  {
+    "ariaLabel": "TTY: 7 1 1.",
+    "column": 4,
+    "href": "tel:+1711",
+    "order": 4,
+    "target": null,
+    "title": "TTY: 711"
+  },
+  {
+    "column": 4,
+    "label": "Visit a medical center or regional office",
+    "href": "https://www.va.gov/find-locations/",
+    "order": 5,
+    "target": null,
+    "title": "Find a VA location"
+  }
+]
