+++ src/lib/drupal/getGlobalElements.ts
@@ -2,10 +2,11 @@ import { formatter } from '@/data/queries/banners'
 import { drupalClient } from '@/lib/drupal/drupalClient'
 import { LayoutProps } from '@/templates/globals/wrapper'
 import { NodeBanner } from '@/types/dataTypes/drupal/node'
+import { queries } from '@/data/queries'
 
-// This is a helper function to fetch global elements for layout.
-// This is going to be run for every pages on build.
-// To make this fast, you could cache the results example on Redis.
+// Helper function to fetch global elements for layout.
+// It is called once for every page on build.
+// TODO: add cache layer to drupalClient query results
 const nonSlugRoute = `${process.env.NEXT_PUBLIC_DRUPAL_BASE_URL}/jsonapi/banner-alerts?item-path=/`
 
 type GlobalElements = LayoutProps
@@ -25,7 +26,12 @@ export async function getGlobalElements(
     await requestBanner.json()
   )
   const banners = formatter(bannerData as NodeBanner[])
+
+  // gather data for header and footer object
+  const headerFooterData = await queries.getData('header-footer-data')
+
   return {
     bannerData: banners,
+    headerFooterData
   }
 }
