+++ src/pages/_document.tsx
@@ -12,11 +12,6 @@ const Document = () => {
       <Head>
         <meta charSet="utf-8" />
         <link
-          rel="stylesheet"
-          data-entry-name="style.css"
-          href="https://unpkg.com/@department-of-veterans-affairs/formation/dist/formation.min.css"
-        />
-        <link
           href={`${ASSETS_URL}static-pages.css`}
           data-entry-name="static-pages.css"
           rel="stylesheet"
@@ -26,13 +21,7 @@ const Document = () => {
           data-entry-name="style.css"
           href={`${ASSETS_URL}style.css`}
         />
-
-        <Script
-          id="headerFooter"
-          strategy="beforeInteractive"
-          src="/importHeaderData.js"
-        />
-
+        {/* Loads widgets built from vets-website after data has been added to window */}
         <Script
           id="staticPages"
           strategy="afterInteractive"
