+++ src/templates/globals/wrapper/index.tsx
@@ -1,3 +1,4 @@
+/* eslint-ignore no-console */
 import { useState, useEffect } from 'react'
 import { Banner } from '@/templates/globals/banners/banner'
 import { PromoBanner } from '@/templates/globals/banners/promoBanner'
@@ -8,11 +9,1042 @@ import { BannerDisplayType, BannerTypeMapping } from '@/data/queries/banners'
 import { Header } from '../header'
 import { Footer } from '../footer/index'
 import { BannerType, FacilityBannerType, PromoBannerType } from '@/types/index'
+
+// Allows additions to window object without overwriting global type
+interface customWindow extends Window {
+  VetsGov?: {
+    headerFooter?: any
+  }
+}
+declare const window: customWindow
+
 export interface LayoutProps {
   children?: React.ReactNode
   bannerData?: Array<
     NodeBanner | PromoBannerType | BannerType | FacilityBannerType
   >
+  // type this
+  headerFooterData?: any
+}
+
+// tmp storage
+const headerData = {
+  "footerData": [
+    {
+      "column": 1,
+      "href": "https://www.va.gov/homeless/",
+      "order": 1,
+      "target": "",
+      "title": "Homeless Veterans"
+    },
+    {
+      "column": 1,
+      "href": "https://www.va.gov/womenvet/",
+      "order": 2,
+      "target": "",
+      "title": "Women Veterans"
+    },
+    {
+      "column": 1,
+      "href": "https://www.va.gov/centerforminorityveterans/",
+      "order": 3,
+      "target": "",
+      "title": "Minority Veterans"
+    },
+    {
+      "column": 1,
+      "href": "https://www.patientcare.va.gov/lgbt/",
+      "order": 4,
+      "target": "",
+      "title": "LGBTQ+ Veterans",
+      "rel": "noopener noreferrer"
+    },
+    {
+      "column": 1,
+      "href": "https://www.ptsd.va.gov",
+      "order": 5,
+      "target": "",
+      "title": "PTSD",
+      "rel": "noopener noreferrer"
+    },
+    {
+      "column": 1,
+      "href": "https://www.mentalhealth.va.gov",
+      "order": 6,
+      "target": "",
+      "title": "Mental health",
+      "rel": "noopener noreferrer"
+    },
+    {
+      "column": 1,
+      "href": "https://www.va.gov/adaptivesports/",
+      "order": 7,
+      "target": "",
+      "title": "Adaptive sports and special events"
+    },
+    {
+      "column": 1,
+      "href": "https://www.va.gov/outreach-and-events/events/",
+      "order": 8,
+      "target": null,
+      "title": "VA outreach events"
+    },
+    {
+      "column": 1,
+      "href": "https://www.nrd.gov",
+      "order": 9,
+      "target": "_blank",
+      "title": "National Resource Directory",
+      "rel": "noopener noreferrer"
+    },
+    {
+      "column": 2,
+      "href": "https://www.va.gov/find-forms/",
+      "order": 1,
+      "target": null,
+      "title": "VA forms"
+    },
+    {
+      "column": 2,
+      "href": "https://www.accesstocare.va.gov/",
+      "order": 2,
+      "target": "_blank",
+      "title": "VA health care access and quality",
+      "rel": "noopener noreferrer"
+    },
+    {
+      "column": 2,
+      "href": "https://www.va.gov/ogc/accreditation.asp",
+      "order": 3,
+      "target": null,
+      "title": "Accredited claims representatives"
+    },
+    {
+      "column": 2,
+      "href": "https://www.mobile.va.gov/appstore/",
+      "order": 4,
+      "target": "_blank",
+      "title": "VA mobile apps",
+      "rel": "noopener noreferrer"
+    },
+    {
+      "column": 2,
+      "href": "https://www.va.gov/statedva.htm",
+      "order": 5,
+      "target": null,
+      "title": "State Veterans Affairs offices"
+    },
+    {
+      "column": 2,
+      "href": "https://www.va.gov/landing2_business.htm",
+      "order": 6,
+      "target": null,
+      "title": "Doing business with VA"
+    },
+    {
+      "column": 2,
+      "href": "https://www.va.gov/jobs/",
+      "order": 7,
+      "target": null,
+      "title": "Careers at VA"
+    },
+    {
+      "column": 2,
+      "href": "https://www.va.gov/outreach-and-events/outreach-materials/",
+      "order": 8,
+      "target": null,
+      "title": "VA outreach materials"
+    },
+    {
+      "column": 2,
+      "href": "https://www.va.gov/welcome-kit/",
+      "order": 9,
+      "target": null,
+      "title": "Your VA welcome kit"
+    },
+    {
+      "column": 3,
+      "href": "https://www.news.va.gov/",
+      "order": 1,
+      "target": "_blank",
+      "title": "VA news",
+      "rel": "noopener noreferrer"
+    },
+    {
+      "column": 3,
+      "href": "https://www.va.gov/opa/pressrel/",
+      "order": 2,
+      "target": null,
+      "title": "Press releases"
+    },
+    {
+      "column": 3,
+      "href": "https://public.govdelivery.com/accounts/USVA/subscriber/new/",
+      "order": 3,
+      "target": "_blank",
+      "title": "Email updates",
+      "rel": "noopener noreferrer"
+    },
+    {
+      "column": 3,
+      "href": "https://www.facebook.com/VeteransAffairs",
+      "order": 4,
+      "target": "_blank",
+      "title": "Facebook",
+      "rel": "noopener noreferrer"
+    },
+    {
+      "column": 3,
+      "href": "https://www.instagram.com/deptvetaffairs/",
+      "order": 5,
+      "target": "_blank",
+      "title": "Instagram",
+      "rel": "noopener noreferrer"
+    },
+    {
+      "column": 3,
+      "href": "https://www.twitter.com/DeptVetAffairs/",
+      "order": 6,
+      "target": "_blank",
+      "title": "Twitter",
+      "rel": "noopener noreferrer"
+    },
+    {
+      "column": 3,
+      "href": "https://www.flickr.com/photos/VeteransAffairs/",
+      "order": 7,
+      "target": "_blank",
+      "title": "Flickr",
+      "rel": "noopener noreferrer"
+    },
+    {
+      "column": 3,
+      "href": "https://www.youtube.com/user/DeptVetAffairs",
+      "order": 8,
+      "target": "_blank",
+      "title": "YouTube",
+      "rel": "noopener noreferrer"
+    },
+    {
+      "column": 3,
+      "href": "https://www.va.gov/opa/socialmedia.asp",
+      "order": 9,
+      "target": null,
+      "title": "All VA social media"
+    },
+    {
+      "column": 4,
+      "href": "https://www.va.gov/resources/",
+      "order": 1,
+      "target": null,
+      "title": "Resources and support"
+    },
+    {
+      "column": 4,
+      "href": "https://www.va.gov/contact-us/",
+      "order": 2,
+      "target": null,
+      "title": "Contact us"
+    },
+    {
+      "column": 4,
+      "label": "Call us",
+      "href": "tel:18006982411",
+      "order": 3,
+      "target": null,
+      "title": "800-698-2411"
+    },
+    {
+      "ariaLabel": "TTY: 7 1 1.",
+      "column": 4,
+      "href": "tel:+1711",
+      "order": 4,
+      "target": null,
+      "title": "TTY: 711"
+    },
+    {
+      "column": 4,
+      "label": "Visit a medical center or regional office",
+      "href": "https://www.va.gov/find-locations/",
+      "order": 5,
+      "target": null,
+      "title": "Find a VA location"
+    },
+    {
+      "column": "bottom_rail",
+      "href": "https://www.va.gov/accessibility-at-va",
+      "order": 1,
+      "target": null,
+      "title": "Accessibility",
+      "rel": "noopener noreferrer"
+    },
+    {
+      "column": "bottom_rail",
+      "href": "https://www.va.gov/resources/your-civil-rights-and-how-to-file-a-discrimination-complaint/",
+      "order": 2,
+      "target": null,
+      "title": "Civil Rights"
+    },
+    {
+      "column": "bottom_rail",
+      "href": "https://www.va.gov/foia/",
+      "order": 3,
+      "target": null,
+      "title": "Freedom of Information Act (FOIA)"
+    },
+    {
+      "column": "bottom_rail",
+      "href": "https://www.va.gov/oig/",
+      "order": 4,
+      "target": null,
+      "title": "Office of Inspector General"
+    },
+    {
+      "column": "bottom_rail",
+      "href": "https://www.va.gov/opa/Plain_Language.asp",
+      "order": 5,
+      "target": null,
+      "title": "Plain language"
+    },
+    {
+      "column": "bottom_rail",
+      "href": "https://www.va.gov/privacy-policy/",
+      "order": 6,
+      "target": null,
+      "title": "Privacy, policies, and legal information"
+    },
+    {
+      "column": "bottom_rail",
+      "href": "https://www.va.gov/privacy/",
+      "order": 7,
+      "target": null,
+      "title": "VA Privacy Service"
+    },
+    {
+      "column": "bottom_rail",
+      "href": "https://www.va.gov/ormdi/NOFEAR_Select.asp",
+      "order": 8,
+      "target": null,
+      "title": "No FEAR Act Data"
+    },
+    {
+      "column": "bottom_rail",
+      "href": "https://www.usa.gov/",
+      "order": 9,
+      "target": "_blank",
+      "title": "USA.gov",
+      "rel": "noopener noreferrer"
+    },
+    {
+      "column": "bottom_rail",
+      "href": "https://www.va.gov/veterans-portrait-project/",
+      "order": 10,
+      "target": null,
+      "title": "Veterans Portrait Project"
+    }
+  ],
+  "megaMenuData": [
+    {
+      "title": "VA benefits and health care",
+      "menuSections": [
+        {
+          "title": "Health care",
+          "links": {
+            "seeAllLink": {
+              "text": "View all in health care",
+              "href": "https://www.va.gov/health-care"
+            },
+            "columnOne": {
+              "title": "Get health care benefits",
+              "links": [
+                {
+                  "text": "About VA health benefits",
+                  "href": "https://www.va.gov/health-care/about-va-health-benefits"
+                },
+                {
+                  "text": "How to apply",
+                  "href": "https://www.va.gov/health-care/how-to-apply"
+                },
+                {
+                  "text": "Family and caregiver health benefits",
+                  "href": "https://www.va.gov/health-care/family-caregiver-benefits"
+                },
+                {
+                  "text": "Apply now for health care",
+                  "href": "https://www.va.gov/health-care/apply/application"
+                }
+              ]
+            },
+            "columnTwo": {
+              "title": "Manage your health",
+              "links": [
+                {
+                  "text": "Refill and track your prescriptions",
+                  "href": "https://www.va.gov/health-care/refill-track-prescriptions"
+                },
+                {
+                  "text": "Send a secure message to your health care team",
+                  "href": "https://www.va.gov/health-care/secure-messaging"
+                },
+                {
+                  "text": "Schedule and manage health appointments",
+                  "href": "https://www.va.gov/health-care/schedule-view-va-appointments"
+                },
+                {
+                  "text": "View your lab and test results",
+                  "href": "https://www.va.gov/health-care/view-test-and-lab-results"
+                },
+                {
+                  "text": "Order hearing aid batteries and prosthetic socks",
+                  "href": "https://www.va.gov/health-care/order-hearing-aid-batteries-and-accessories"
+                }
+              ]
+            },
+            "columnThree": {
+              "img": {
+                "src": "https://www.va.gov/img/styles/3_2_medium_thumbnail/public/hub_promos/health-care.png",
+                "alt": ""
+              },
+              "link": {
+                "text": "Coronavirus FAQs",
+                "href": "https://www.va.gov/coronavirus-veteran-frequently-asked-questions/"
+              },
+              "description": "Get answers to frequently asked questions about COVID-19 and VA health care. You can also call our MyVA411 main information line at at 800-698-2411."
+            }
+          }
+        },
+        {
+          "title": "Disability",
+          "links": {
+            "seeAllLink": {
+              "text": "View all in disability",
+              "href": "https://www.va.gov/disability"
+            },
+            "columnOne": {
+              "title": "Get disability benefits",
+              "links": [
+                {
+                  "text": "Eligibility",
+                  "href": "https://www.va.gov/disability/eligibility"
+                },
+                {
+                  "text": "How to file a claim",
+                  "href": "https://www.va.gov/disability/how-to-file-claim"
+                },
+                {
+                  "text": "Survivor and dependent compensation (DIC)",
+                  "href": "https://www.va.gov/disability/dependency-indemnity-compensation"
+                }
+              ]
+            },
+            "columnTwo": {
+              "title": "Manage your benefits",
+              "links": [
+                {
+                  "text": "Check your claim or appeal status",
+                  "href": "https://www.va.gov/claim-or-appeal-status"
+                },
+                {
+                  "text": "View your VA payment history",
+                  "href": "https://www.va.gov/va-payment-history"
+                },
+                {
+                  "text": "Upload evidence to support your claim",
+                  "href": "https://www.va.gov/disability/upload-supporting-evidence"
+                },
+                {
+                  "text": "File for a VA disability increase",
+                  "href": "https://www.va.gov/disability/how-to-file-claim"
+                },
+                {
+                  "text": "Decision reviews and appeals",
+                  "href": "https://www.va.gov/decision-reviews"
+                }
+              ]
+            },
+            "columnThree": {
+              "img": {
+                "src": "https://www.va.gov/img/styles/3_2_medium_thumbnail/public/hub_promos/disability.png",
+                "alt": ""
+              },
+              "link": {
+                "text": "Fully Developed Claims program",
+                "href": "https://www.va.gov/disability/how-to-file-claim/evidence-needed/fully-developed-claims"
+              },
+              "description": "Use this program to get a faster decision by submitting evidence along with your claim."
+            }
+          }
+        },
+        {
+          "title": "Education and training",
+          "links": {
+            "columnOne": {
+              "title": "Get education benefits",
+              "links": [
+                {
+                  "text": "About GI Bill benefits",
+                  "href": "https://www.va.gov/education/about-gi-bill-benefits"
+                },
+                {
+                  "text": "Eligibility",
+                  "href": "https://www.va.gov/education/eligibility"
+                },
+                {
+                  "text": "How to apply",
+                  "href": "https://www.va.gov/education/how-to-apply"
+                },
+                {
+                  "text": "Veteran Readiness and Employment",
+                  "href": "https://www.va.gov/careers-employment/vocational-rehabilitation"
+                },
+                {
+                  "text": "Survivor and dependent education benefits",
+                  "href": "https://www.va.gov/education/survivor-dependent-benefits"
+                }
+              ]
+            },
+            "columnTwo": {
+              "title": "Manage your benefits",
+              "links": [
+                {
+                  "text": "View your VA payment history",
+                  "href": "https://www.va.gov/va-payment-history"
+                },
+                {
+                  "text": "Check your Post-9/11 GI Bill benefits",
+                  "href": "https://www.va.gov/education/gi-bill/post-9-11/ch-33-benefit/"
+                },
+                {
+                  "text": "Transfer your Post-9/11 GI Bill benefits",
+                  "href": "https://www.va.gov/education/transfer-post-9-11-gi-bill-benefits"
+                },
+                {
+                  "text": "Change your GI Bill school or program",
+                  "href": "https://www.va.gov/education/change-gi-bill-benefits"
+                },
+                {
+                  "text": "Change your direct deposit information",
+                  "href": "https://www.va.gov/change-direct-deposit"
+                }
+              ]
+            },
+            "columnThree": {
+              "img": {
+                "src": "https://www.va.gov/img/styles/3_2_medium_thumbnail/public/hub_promos/education.png",
+                "alt": ""
+              },
+              "link": {
+                "text": "GI Bill® Comparison Tool",
+                "href": "https://www.va.gov/education/gi-bill-comparison-tool"
+              },
+              "description": "Learn about education programs and compare benefits by school."
+            },
+            "seeAllLink": {
+              "text": "View all in education",
+              "href": "https://www.va.gov/education"
+            }
+          }
+        },
+        {
+          "title": "Careers and employment",
+          "links": {
+            "seeAllLink": {
+              "text": "View all in careers and employment",
+              "href": "https://www.va.gov/careers-employment"
+            },
+            "columnOne": {
+              "title": "Get employment benefits",
+              "links": [
+                {
+                  "text": "About Veteran Readiness and Employment",
+                  "href": "https://www.va.gov/careers-employment/vocational-rehabilitation"
+                },
+                {
+                  "text": "How to apply",
+                  "href": "https://www.va.gov/careers-employment/vocational-rehabilitation/how-to-apply"
+                },
+                {
+                  "text": "Veteran-owned small business support",
+                  "href": "https://www.va.gov/careers-employment/veteran-owned-business-support"
+                }
+              ]
+            },
+            "columnTwo": {
+              "title": "Manage your career",
+              "links": [
+                {
+                  "text": "VA transition assistance",
+                  "href": "https://www.benefits.va.gov/tap/"
+                },
+                {
+                  "text": "CareerScope assessment",
+                  "href": "https://www.va.gov/careers-employment/careerscope-skills-assessment"
+                },
+                {
+                  "text": "Find a job",
+                  "href": "https://www.dol.gov/veterans/findajob/"
+                },
+                {
+                  "text": "Find VA careers and support",
+                  "href": "https://www.va.gov/jobs/"
+                },
+                {
+                  "text": "Print your civil service preference letter",
+                  "href": "https://www.va.gov/records/download-va-letters"
+                }
+              ]
+            },
+            "columnThree": {
+              "img": {
+                "src": "https://www.va.gov/img/styles/3_2_medium_thumbnail/public/hub_promos/careers.png",
+                "alt": ""
+              },
+              "link": {
+                "text": "VA for Vets",
+                "href": "https://www.vaforvets.va.gov/"
+              },
+              "description": "Get help transitioning to a civilian career and matching your skills and experiences to VA job opportunities."
+            }
+          }
+        },
+        {
+          "title": "Pension",
+          "links": {
+            "seeAllLink": {
+              "text": "View all in pension",
+              "href": "https://www.va.gov/pension"
+            },
+            "columnOne": {
+              "title": "Get pension benefits",
+              "links": [
+                {
+                  "text": "Veterans Pension eligibility",
+                  "href": "https://www.va.gov/pension/eligibility"
+                },
+                {
+                  "text": "How to apply",
+                  "href": "https://www.va.gov/pension/how-to-apply"
+                },
+                {
+                  "text": "Aid and attendance benefits and housebound allowance",
+                  "href": "https://www.va.gov/pension/aid-attendance-housebound"
+                },
+                {
+                  "text": "Survivors Pension",
+                  "href": "https://www.va.gov/pension/survivors-pension"
+                },
+                {
+                  "text": "Apply now for a Veterans Pension",
+                  "href": "https://www.va.gov/pension/application/527EZ"
+                }
+              ]
+            },
+            "columnTwo": {
+              "title": "Manage your benefits",
+              "links": [
+                {
+                  "text": "Check your claim or appeal status",
+                  "href": "https://www.va.gov/claim-or-appeal-status"
+                },
+                {
+                  "text": "View your VA payment history",
+                  "href": "https://www.va.gov/va-payment-history"
+                },
+                {
+                  "text": "Change your address",
+                  "href": "https://www.va.gov/change-address"
+                },
+                {
+                  "text": "Change your direct deposit information",
+                  "href": "https://www.va.gov/change-direct-deposit"
+                }
+              ]
+            },
+            "columnThree": {
+              "img": {
+                "src": "https://www.va.gov/img/styles/3_2_medium_thumbnail/public/hub_promos/pension.png",
+                "alt": ""
+              },
+              "link": {
+                "text": "Get your pension payments electronically",
+                "href": "https://www.godirect.gov/gpw"
+              },
+              "description": "Sign up for direct deposit or a pre-paid debit card."
+            }
+          }
+        },
+        {
+          "title": "Housing assistance",
+          "links": {
+            "seeAllLink": {
+              "text": "View all in housing assistance",
+              "href": "https://www.va.gov/housing-assistance"
+            },
+            "columnOne": {
+              "title": "Get home loan benefits",
+              "links": [
+                {
+                  "text": "About VA home loan types",
+                  "href": "https://www.va.gov/housing-assistance/home-loans/loan-types"
+                },
+                {
+                  "text": "Check your appeal status",
+                  "href": "https://www.va.gov/claim-or-appeal-status"
+                },
+                {
+                  "text": "How to apply for your COE",
+                  "href": "https://www.va.gov/housing-assistance/home-loans/how-to-apply"
+                }
+              ]
+            },
+            "columnTwo": {
+              "title": "Get Veterans housing grants",
+              "links": [
+                {
+                  "text": "About disability housing grants",
+                  "href": "https://www.va.gov/housing-assistance/disability-housing-grants"
+                },
+                {
+                  "text": "Check your SAH claim status",
+                  "href": "https://www.va.gov/claim-or-appeal-status"
+                },
+                {
+                  "text": "How to apply for an SAH grant",
+                  "href": "https://www.va.gov/housing-assistance/disability-housing-grants/how-to-apply"
+                }
+              ]
+            },
+            "columnThree": {
+              "img": {
+                "src": "https://www.va.gov/img/styles/3_2_medium_thumbnail/public/hub_promos/housing.png",
+                "alt": ""
+              },
+              "link": {
+                "text": "National Call Center for Homeless Veterans",
+                "href": "https://www.va.gov/homeless/nationalcallcenter.asp"
+              },
+              "description": "Find out how we can help you or a Veteran you care about connect with VA and local resources 24 hours a day, 7 days a week."
+            }
+          }
+        },
+        {
+          "title": "Life insurance",
+          "links": {
+            "seeAllLink": {
+              "text": "View all in life insurance",
+              "href": "https://www.va.gov/life-insurance"
+            },
+            "columnOne": {
+              "title": "Get life insurance",
+              "links": [
+                {
+                  "text": "About life insurance options",
+                  "href": "https://www.va.gov/life-insurance/options-eligibility"
+                },
+                {
+                  "text": "Benefits for totally disabled or terminally ill policyholders",
+                  "href": "https://www.va.gov/life-insurance/totally-disabled-or-terminally-ill"
+                },
+                {
+                  "text": "Beneficiary financial counseling and online will preparation",
+                  "href": "https://www.benefits.va.gov/insurance/bfcs.asp"
+                }
+              ]
+            },
+            "columnTwo": {
+              "title": "Manage your life insurance",
+              "links": [
+                {
+                  "text": "Access your policy online",
+                  "href": "https://www.va.gov/life-insurance/manage-your-policy"
+                },
+                {
+                  "text": "Update your beneficiaries",
+                  "href": "https://www.benefits.va.gov/INSURANCE/updatebene.asp"
+                },
+                {
+                  "text": "File a claim for insurance benefits",
+                  "href": "https://www.benefits.va.gov/INSURANCE/sglivgli.asp"
+                },
+                {
+                  "text": "Check your appeal status",
+                  "href": "https://www.va.gov/claim-or-appeal-status"
+                }
+              ]
+            },
+            "columnThree": {
+              "img": {
+                "src": "https://www.va.gov/img/styles/3_2_medium_thumbnail/public/hub_promos/life-insurance.png",
+                "alt": ""
+              },
+              "link": {
+                "text": "SGLI Online Enrollment System (SOES)",
+                "href": "https://www.benefits.va.gov/INSURANCE/SOES.asp"
+              },
+              "description": "Learn about our new online enrollment system for Servicemembers' Group Life Insurance."
+            }
+          }
+        },
+        {
+          "title": "Burials and memorials",
+          "links": {
+            "columnOne": {
+              "title": "Get burial benefits",
+              "links": [
+                {
+                  "text": "Eligibility",
+                  "href": "https://www.va.gov/burials-memorials/eligibility"
+                },
+                {
+                  "text": "Pre-need burial eligibility determination",
+                  "href": "https://www.va.gov/burials-memorials/pre-need-eligibility"
+                },
+                {
+                  "text": "Veteran burial allowance",
+                  "href": "https://www.va.gov/burials-memorials/veterans-burial-allowance"
+                },
+                {
+                  "text": "Memorial items",
+                  "href": "https://www.va.gov/burials-memorials/memorial-items"
+                },
+                {
+                  "text": "Survivor and dependent compensation (DIC)",
+                  "href": "https://www.va.gov/disability/dependency-indemnity-compensation"
+                }
+              ]
+            },
+            "columnTwo": {
+              "title": "Plan a burial",
+              "links": [
+                {
+                  "text": "Schedule a burial for a family member",
+                  "href": "https://www.va.gov/burials-memorials/schedule-a-burial"
+                },
+                {
+                  "text": "Schedule a burial in a VA national cemetery",
+                  "href": "https://www.cem.va.gov/cem/burial_benefits/need.asp"
+                },
+                {
+                  "text": "Find a cemetery",
+                  "href": "https://www.cem.va.gov/cems/listcem.asp"
+                },
+                {
+                  "text": "Request military records (DD214)",
+                  "href": "https://www.va.gov/records/get-military-service-records"
+                }
+              ]
+            },
+            "seeAllLink": {
+              "text": "View all in burials and memorials",
+              "href": "https://www.va.gov/burials-memorials"
+            },
+            "columnThree": {
+              "img": {
+                "src": "https://www.va.gov/img/styles/3_2_medium_thumbnail/public/hub_promos/burials.png",
+                "alt": ""
+              },
+              "link": {
+                "text": "Find National Cemetery Administration Information",
+                "href": "https://m.va.gov/"
+              },
+              "description": "Locate a grave, search for cemeteries, and find benefits information and resources."
+            }
+          }
+        },
+        {
+          "title": "Records",
+          "links": {
+            "seeAllLink": {
+              "text": "View all in records",
+              "href": "https://www.va.gov/records"
+            },
+            "columnOne": {
+              "title": "Get your records",
+              "links": [
+                {
+                  "text": "Get your VA medical records (Blue Button)",
+                  "href": "https://www.va.gov/health-care/get-medical-records"
+                },
+                {
+                  "text": "Download your VA benefits letters",
+                  "href": "https://www.va.gov/records/download-va-letters"
+                },
+                {
+                  "text": "Learn how to apply for a home loan COE",
+                  "href": "https://www.va.gov/housing-assistance/home-loans/how-to-apply"
+                },
+                {
+                  "text": "Get Veteran ID cards",
+                  "href": "https://www.va.gov/records/get-veteran-id-cards"
+                }
+              ]
+            },
+            "columnTwo": {
+              "title": "Manage your records",
+              "links": [
+                {
+                  "text": "Request your military records (DD214)",
+                  "href": "https://www.va.gov/records/get-military-service-records"
+                },
+                {
+                  "text": "Change your address",
+                  "href": "https://www.va.gov/change-address"
+                },
+                {
+                  "text": "How to apply for a discharge upgrade",
+                  "href": "https://www.va.gov/discharge-upgrade-instructions"
+                },
+                {
+                  "text": "View your VA payment history",
+                  "href": "https://www.va.gov/va-payment-history"
+                },
+                {
+                  "text": "Search historical military records (National Archives)",
+                  "href": "https://www.archives.gov/"
+                }
+              ]
+            },
+            "columnThree": {
+              "img": {
+                "src": "https://www.va.gov/img/styles/3_2_medium_thumbnail/public/hub_promos/records.png",
+                "alt": ""
+              },
+              "link": {
+                "text": "Confirm your VA benefit status",
+                "href": "https://www.va.gov/records/download-va-letters"
+              },
+              "description": "Download letters like your eligibility or award letter for certain benefits."
+            }
+          }
+        },
+        {
+          "title": "Service member benefits",
+          "text": "Service member benefits",
+          "href": "https://www.va.gov/service-member-benefits"
+        },
+        {
+          "title": "Family member benefits",
+          "text": "Family member benefits",
+          "href": "https://www.va.gov/family-member-benefits"
+        }
+      ]
+    },
+    {
+      "title": "About VA",
+      "menuSections": {
+        "mainColumn": {
+          "title": "VA organizations",
+          "links": [
+            {
+              "text": "Veterans Health Administration",
+              "href": "https://www.va.gov/health"
+            },
+            {
+              "text": "Veterans Benefits Administration",
+              "href": "https://www.benefits.va.gov/benefits/"
+            },
+            {
+              "text": "National Cemetery Administration",
+              "href": "https://www.cem.va.gov/"
+            },
+            {
+              "text": "VA Leadership",
+              "href": "https://www.va.gov/opa/bios/index.asp"
+            },
+            {
+              "text": "Public Affairs",
+              "href": "https://www.va.gov/OPA/index.asp"
+            },
+            {
+              "text": "Congressional Affairs",
+              "href": "https://www.va.gov/oca/index.asp"
+            },
+            {
+              "text": "All VA offices and organizations",
+              "href": "https://www.va.gov/landing_organizations.htm"
+            }
+          ]
+        },
+        "columnOne": {
+          "title": "Innovation at VA",
+          "links": [
+            {
+              "text": "Health research",
+              "href": "https://www.research.va.gov/"
+            },
+            {
+              "text": "Public health",
+              "href": "https://www.publichealth.va.gov/"
+            },
+            {
+              "text": "VA open data",
+              "href": "https://www.va.gov/data/"
+            },
+            {
+              "text": "Veterans analysis and statistics",
+              "href": "https://www.va.gov/VETDATA/index.asp"
+            },
+            {
+              "text": "Appeals modernization",
+              "href": "https://www.benefits.va.gov/benefits/appeals.asp"
+            },
+            {
+              "text": "VA Innovation Center",
+              "href": "https://www.innovation.va.gov/"
+            },
+            {
+              "text": "Recovery Act",
+              "href": "https://www.va.gov/recovery/"
+            }
+          ]
+        },
+        "columnTwo": {
+          "title": "Learn about VA",
+          "links": [
+            {
+              "text": "About VA",
+              "href": "https://www.va.gov/ABOUT_VA/index.asp"
+            },
+            {
+              "text": "History of VA",
+              "href": "https://www.va.gov/about_va/vahistory.asp"
+            },
+            {
+              "text": "VA plans, budget, finances, and performance",
+              "href": "https://www.va.gov/performance/"
+            },
+            {
+              "text": "National cemetery history program",
+              "href": "https://www.cem.va.gov/cem/history/index.asp"
+            },
+            {
+              "text": "Veterans legacy program",
+              "href": "https://www.cem.va.gov/cem/legacy/index.asp"
+            },
+            {
+              "text": "Volunteer or donate",
+              "href": "https://www.volunteer.va.gov/index.asp"
+            }
+          ]
+        },
+        "columnThree": {
+          "img": {
+            "src": "https://www.va.gov/img/styles/3_2_medium_thumbnail/public/2019-11/afr-2019_0.jpg",
+            "alt": "Agency Financial Report 2019"
+          },
+          "link": {
+            "text": "Agency Financial Report",
+            "href": "https://www.va.gov/finance/afr/index.asp"
+          },
+          "description": "View the 2019 report on VA’s progress in providing benefits and health care."
+        }
+      }
+    },
+    {
+      "title": "Find a VA location",
+      "href": "https://www.va.gov/find-locations"
+    }
+  ]
 }
 
 export const formatBannerType = (bannerData) => {
@@ -28,10 +1060,18 @@ export const formatBannerType = (bannerData) => {
   }
 }
 
-export function Wrapper({ bannerData, children }: LayoutProps) {
+export function Wrapper({ bannerData, headerFooterData, children }: LayoutProps) {
   const [showBanners, setShowBanners] = useState(false)
   const [banners, setBanners] = useState([])
   useEffect(() => {
+    // Place header & footer data on window object for vets-website widgets
+    window.VetsGov = {}
+    window.VetsGov.headerFooter = headerData
+
+    console.log(headerData)
+    console.log(headerFooterData)
+
+    // todo: clean this up later
     if (isEmpty(bannerData)) {
       return setShowBanners(false)
     } else {
